PORTALS-3407
- Refactor JSON Schema form components and templates for improved usability, focused on decoupling generic UI from annotations logic
- Refactor to extract a generic StyledPopover
- Add feature toggles to JsonSchemaForm:
  - descriptionVariant: toggle between popover (new) and expand/collapse
  - descriptionFormat: toggle between string (new) and table
  - showDerivedAnnotationPlaceholder: opt-in to show placeholders for derived values (annotations only)
  - allowFreeSoloEnum: opt-in to allow free text for enum fields (annotations only)
 - Upgrade RJSF to v5.24.8
