+++ packages/synapse-react-client/package.json
@@ -48,10 +48,10 @@
     "@react-google-maps/api": "^2.19.3",
     "@react-hook/resize-observer": "^1.2.6",
     "@react-hookz/web": "^23.1.0",
-    "@rjsf/core": "5.17.1",
-    "@rjsf/mui": "5.17.1",
-    "@rjsf/utils": "5.17.1",
-    "@rjsf/validator-ajv8": "5.17.1",
+    "@rjsf/core": "5.24.8",
+    "@rjsf/mui": "5.24.8",
+    "@rjsf/utils": "5.24.8",
+    "@rjsf/validator-ajv8": "5.24.8",
     "@sage-bionetworks/markdown-it-container": "workspace:*",
     "@sage-bionetworks/synapse-client": "workspace:*",
     "@sage-bionetworks/synapse-types": "workspace:*",
