+++ packages/synapse-react-client/src/components/JsonSchemaForm/templates/ArrayFieldDescriptionTemplate.tsx
