+++ packages/synapse-react-client/src/components/JsonSchemaForm/templates/ArrayFieldItemTemplate.tsx
