+++ packages/synapse-react-client/src/components/JsonSchemaForm/templates/ArrayFieldTitleTemplate.tsx
