+++ packages/synapse-react-client/src/components/JsonSchemaForm/templates/DescriptionFieldTemplate.tsx
@@ -1,9 +1,13 @@
+import { CustomFormContext } from '@/components/JsonSchemaForm/CustomFormContext'
 import { DescriptionFieldProps } from '@rjsf/utils'
 import { Typography } from '@mui/material'
 import FieldDescriptionTable from './FieldDescriptionTable'
 
 export default function DescriptionFieldTemplate(props: DescriptionFieldProps) {
-  const { description, schema = {} } = props
+  const { description, schema = {}, registry } = props
+
+  const descriptionFormat =
+    (registry.formContext as CustomFormContext).descriptionFormat || 'string'
 
   if (schema.type === 'object') {
     return (
@@ -11,6 +15,10 @@ export default function DescriptionFieldTemplate(props: DescriptionFieldProps) {
         {description}
       </Typography>
     )
+  }
+
+  if (descriptionFormat === 'string') {
+    return <Typography variant={'body1'}>{description}</Typography>
   }
 
   return <FieldDescriptionTable schema={schema} description={description} />
