+++ packages/synapse-react-client/src/components/JsonSchemaForm/templates/FieldDescriptionTable.tsx
