+++ packages/synapse-react-client/src/components/JsonSchemaForm/widgets/BooleanWidget.tsx
