+++ packages/synapse-react-client/src/components/SchemaDrivenAnnotationEditor/template/AdditionalPropertyContext.tsx
@@ -8,7 +8,7 @@ export type AdditionalPropertyContextType = {
  * The `AdditionalPropertyContext` is a custom context object used to pass the `dropProperty` handler to the items within
  * a particular additional property.
  *
- * In RJSF, dropping an entire additional property is handled at the `object` level. In this Synapse UI,
+ * In RJSF, dropping an entire additional property is handled at the `object` level. In the Synapse Annotations UI,
  * we treat all additional properties (a.k.a. custom annotations) as arrays, and render the delete control at the `array`
  * level. RJSF does not normally provide the drop property handler at the `array` level, so we capture it and pass it
  * along using this context.
